web - moved all policy pages to page/policies/ dir
