+++ web/src/pages/policies/terms.tsx
